Move agents from interaction/read to read and interaction/write to write
