+++ db/common/agent/write/ParentshipAgent.java
@@ -1,4 +1,4 @@
-package grakn.simulation.db.common.agent.interaction.write;
+package grakn.simulation.db.common.agent.write;
 
 import grakn.simulation.db.common.action.SpouseType;
 import grakn.simulation.db.common.agent.base.SimulationContext;
